Isolate development mode certificate manager, fix tests
