+++ api/pom.xml
@@ -44,6 +44,9 @@
         <!-- Sonar settings -->
         <sonar.host.url>https://sonarcloud.io</sonar.host.url>
         <sonar.organization>eyefloaters</sonar.organization>
+        <sonar.exclusions>
+        	src/main/java/com/github/eyefloaters/console/api/support/TrustAllCertificateManager.java
+        </sonar.exclusions>
     </properties>
 
     <scm>
@@ -432,6 +435,9 @@
                             <goal>prepare-agent</goal>
                         </goals>
                         <configuration>
+                            <excludes>
+                                <exclude>com/github/eyefloaters/console/api/support/TrustAllCertificateManager.class</exclude>
+                            </excludes>
                             <exclClassLoaders>*QuarkusClassLoader</exclClassLoaders>
                             <destFile>${project.build.directory}/jacoco-quarkus.exec</destFile>
                             <append>true</append>
@@ -443,6 +449,9 @@
                             <goal>prepare-agent-integration</goal>
                         </goals>
                         <configuration>
+                            <excludes>
+                                <exclude>com/github/eyefloaters/console/api/support/TrustAllCertificateManager.class</exclude>
+                            </excludes>
                             <exclClassLoaders>*QuarkusClassLoader</exclClassLoaders>
                             <destFile>${project.build.directory}/jacoco-quarkus.exec</destFile>
                             <append>true</append>
