+++ api/src/main/resources/application.properties
@@ -64,6 +64,8 @@ quarkus.arc.unremovable-types=com.github.eyefloaters.console.api.**
 
 kafka.admin.request.timeout.ms=10000
 kafka.admin.default.api.timeout.ms=10000
+kafka.admin.sasl.login.refresh.min.period.seconds=60
+kafka.admin.sasl.login.refresh.buffer.seconds=60
 
 ########
 #%dev.quarkus.http.auth.proactive=false
